add LLVMCompat for llvm versioning
