+++ CMakeLists.txt
@@ -8,23 +8,33 @@ set (CMAKE_CXX_STANDARD 11)
 find_package(ZLIB REQUIRED)
 include_directories(${ZLIB_INCLUDE_DIRS})
 
-find_package(LLVM 8.0 REQUIRED CONFIG)
+find_package(LLVM REQUIRED CONFIG)
 include_directories(${LLVM_INCLUDE_DIRS})
 add_definitions(${LLVM_DEFINITIONS})
-message(STATUS "Found LLVM ${LLVM_PACKAGE_VERSION}")
+
+string (REGEX REPLACE "^([0-9]+)\.([0-9]+)\.([0-9]+).*" "\\1" LLVM_VERSION_MAJOR "${LLVM_PACKAGE_VERSION}")
+string (REGEX REPLACE "^([0-9]+)\.([0-9]+)\.([0-9]+).*" "\\2" LLVM_VERSION_MINOR "${LLVM_PACKAGE_VERSION}")
+string (REGEX REPLACE "^([0-9]+)\.([0-9]+)\.([0-9]+).*" "\\3" LLVM_VERSION_PATCH "${LLVM_PACKAGE_VERSION}")
+
+message(STATUS "Found  ${LLVM_PACKAGE_VERSION}")
+
+message(STATUS "Found LLVM_VERSION_MAJOR ${LLVM_VERSION_MAJOR}")
+message(STATUS "Found LLVM_VERSION_MINOR ${LLVM_VERSION_MINOR}")
+message(STATUS "Found LLVM_VERSION_PATCH ${LLVM_VERSION_PATCH}")
+
 message(STATUS "Using LLVMConfig.cmake in: ${LLVM_DIR}")
 message(STATUS "LLVM_INCLUDE_DIRS=${LLVM_INCLUDE_DIRS}")
 message(STATUS "LLVM_DEFINITIONS=${LLVM_DEFINITIONS}")
 
 message(STATUS "SYSTEM_PROCESSOR ${CMAKE_SYSTEM_PROCESSOR}")
+
 if (CMAKE_SYSTEM_PROCESSOR MATCHES "(x86)|(X86)|(amd64)|(AMD64)")
   set(llvm_libs_processor
-    x86
-    x86codegen
     x86asmparser
-    x86disassembler
     x86asmprinter
+    x86codegen
     x86desc
+    x86disassembler
     x86info
     x86utils
   )
@@ -81,9 +91,9 @@ function(with_llvm_config trget)
   )
   set(local_libs ${ARGN})
   llvm_map_components_to_libnames(local_llvm_libs ${llvm_libs})
-  message(STATUS "target : ${trget}")
-  message(STATUS "target_libs : ${local_libs}")
-  message(STATUS "llvm_libs : ${local_llvm_libs}")
+  # message(STATUS "target : ${trget}")
+  # message(STATUS "target_libs : ${local_libs}")
+  # message(STATUS "llvm_libs : ${local_llvm_libs}")
 
   get_target_property(t ${trget} TYPE)
   if ("x${t}" STREQUAL "xSTATIC_LIBRARY")
@@ -120,13 +130,13 @@ elseif ((CMAKE_C_COMPILER_ID MATCHES "GNU" OR CMAKE_C_COMPILER_ID MATCHES "Clang
   endif ()
   message(STATUS "Compiling on GNU or Clang platform")
   # set(SANITIZER_FLAGS "-fsanitize=address")
-  set(CMAKE_CXX_FLAGS "-frtti -O2 -fomit-frame-pointer -Wall -Wno-sign-compare -Winline -Wno-delete-non-virtual-dtor -std=c++11 -fexceptions")
+  set(CMAKE_CXX_FLAGS "-frtti -O2 -g3 -fomit-frame-pointer -Wall -Wno-sign-compare -Wno-inline -Wno-delete-non-virtual-dtor -std=c++11 -fexceptions")
   set(CMAKE_CXX_FLAGS_DEBUG "${SANITIZER_FLAGS} -fno-omit-frame-pointer -frtti -O0 -g3 -Wall -Wno-sign-compare -Wno-delete-non-virtual-dtor -std=c++11 -fexceptions")
-  set(CMAKE_CXX_FLAGS_RELEASE "-frtti -DNDEBUG -O2 -fomit-frame-pointer -Wall -Wno-sign-compare -Winline -Wno-delete-non-virtual-dtor -std=c++11 -fexceptions")
+  set(CMAKE_CXX_FLAGS_RELEASE "-frtti -DNDEBUG -O2 -fomit-frame-pointer -Wall -Wno-sign-compare -Wno-inline -Wno-delete-non-virtual-dtor -std=c++11 -fexceptions")
   set(CMAKE_CXX_FLAGS_RELWITHDEBINFO "${SANITIZER_FLAGS} -DNDEBUG -frtti -O2 -g3 -Wall -Wno-sign-compare -Wno-delete-non-virtual-dtor -std=c++11 -fexceptions")
 elseif (APPLE)
   message(STATUS "Compiling on APPLE platform")
-  set(CMAKE_CXX_FLAGS "-O3 -DNDEBUG -Wall -frtti -Wno-sign-compare -Wno-delete-non-virtual-dtor -fexceptions")
+  set(CMAKE_CXX_FLAGS "-O3  -g3 -DNDEBUG -Wall -frtti -Wno-sign-compare -Wno-delete-non-virtual-dtor -fexceptions")
   set(CMAKE_CXX_FLAGS_DEBUG "-frtti -O0 -g3 -Wall -Wno-sign-compare -Wno-delete-non-virtual-dtor -fexceptions")
   set(CMAKE_CXX_FLAGS_RELEASE "-frtti -DNDEBUG -O2 -Wall -Wno-sign-compare -Wno-delete-non-virtual-dtor -fexceptions")
   set(CMAKE_CXX_FLAGS_RELWITHDEBINFO "-frtti -DNDEBUG -O2 -g3 -Wall -Wno-sign-compare -Wno-delete-non-virtual-dtor -fexceptions")
@@ -139,30 +149,42 @@ message(STATUS "CMAKE_CXX_FLAGS_DEBUG=${CMAKE_CXX_FLAGS_DEBUG}")
 message(STATUS "CMAKE_CXX_FLAGS_RELEASE=${CMAKE_CXX_FLAGS_RELEASE}")
 message(STATUS "CMAKE_CXX_FLAGS_RELWITHDEBINFO=${CMAKE_CXX_FLAGS_RELWITHDEBINFO}")
 
+set(llvm_compatibility_dir "${CMAKE_CURRENT_SOURCE_DIR}/LLVMCompat/${LLVM_VERSION_MAJOR}.${LLVM_VERSION_MINOR}.${LLVM_VERSION_PATCH}")
+message(STATUS "${llvm_compatibility_dir}")
+if (NOT IS_DIRECTORY "${llvm_compatibility_dir}")
+  message(FATAL_ERROR "Hobbes lacks ${LLVM_PACKAGE_VERSION} LLVMCompatibility bindings")
+else()
+  message(STATUS "Hobbes has ${LLVM_PACKAGE_VERSION} LLVMCompatibility bindings : [${llvm_compatibility_dir}]")
+  include_directories(${llvm_compatibility_dir}/include)
+endif()
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/LLVMCompat)
+add_subdirectory(${llvm_compatibility_dir})
+
 file(GLOB_RECURSE lib_headers lib/*.H)
 file(GLOB_RECURSE lib_source lib/*.C)
 set(lib_files ${lib_headers} ${lib_source})
 include_directories(include)
 
 add_library(hobbes STATIC ${lib_files})
-with_llvm_config(hobbes ${ZLIB_LIBRARIES})
+with_llvm_config(hobbes llvmCompat ${ZLIB_LIBRARIES})
 
 add_library(hobbes-pic STATIC ${lib_files})
-with_llvm_config(hobbes-pic ${ZLIB_LIBRARIES})
+with_llvm_config(hobbes-pic llvmCompat  ${ZLIB_LIBRARIES})
 set_property(TARGET hobbes-pic PROPERTY POSITION_INDEPENDENT_CODE TRUE)
 
 file(GLOB hi_files bin/hi/*.C)
 add_executable(hi ${hi_files})
-with_llvm_config(hi hobbes ncurses readline ${sys_libs})
+with_llvm_config(hi llvmCompat hobbes ncurses readline ${sys_libs})
 
 file(GLOB_RECURSE hog_files bin/hog/*.C)
 add_executable(hog ${hog_files})
-with_llvm_config(hog hobbes ncurses ${sys_libs})
+with_llvm_config(hog llvmCompat hobbes ncurses ${sys_libs})
 
 enable_testing()
 file(GLOB test_files test/*.C)
 add_executable(hobbes-test ${test_files})
-with_llvm_config(hobbes-test hobbes ncurses ${sys_libs})
+with_llvm_config(hobbes-test llvmCompat hobbes ncurses ${sys_libs})
+
 add_test(hobbes-test hobbes-test)
 include(FindPythonInterp)
 set_property(TARGET hobbes-test PROPERTY COMPILE_FLAGS "-DPYTHON_EXECUTABLE=\"${PYTHON_EXECUTABLE}\" -DSCRIPT_DIR=\"${CMAKE_SOURCE_DIR}/scripts/\"")
